don’t use folder tools on files :-).

+++ app/jobs/ImportJob.java
@@ -56,7 +56,7 @@ public class ImportJob extends AbstractJob {
 			}
 			count--;
 			logger.debug(f.getAbsolutePath()+"!="+MEDIA_FOLDER.getAbsolutePath()+": "+ MediaFileHelper.getCount(f));
-			if(MediaFileHelper.getCount(f) == 0 && !f.getAbsolutePath().equals(MEDIA_FOLDER.getAbsolutePath())) {
+			if(f.isDirectory() && MediaFileHelper.getCount(f) == 0 && !f.getAbsolutePath().equals(MEDIA_FOLDER.getAbsolutePath())) {
 				MediaFileHelper.delete(f.getAbsolutePath().endsWith("/") ? f.getAbsolutePath().substring(0, f.getAbsolutePath().length()-1) : f.getAbsolutePath());
 			}
 		}
